Ich habe es geschafft, dass ich mithilfe der Orginalkoordinaten die Punkte besser zeichnen konnte und Sie in die Forme entsprechend einfügen und skalieren konnte mit allem

+++ DrawingLetters/DrawPoint.cs
@@ -2,15 +2,15 @@
 {
     public class DrawPoint
     {
-        public DrawPoint(float x, float y, int distance = -1)
+        public DrawPoint(double x, double y, int distance = -1)
         {
             X = x;
             Y = y;
             Distance = distance;
         }
 
-        public float X { get; set; }
-        public float Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         public int Distance { get; set; }
 
         public override string ToString() => $"(Distance:{Distance})\n";
